+++ admin/init/model/class_model.php
@@ -356,39 +356,39 @@ class class_model
 
 	public function fetchAll_documentrequest($day = null, $month = null, $year = null)
 	{
-    // Start the base SQL query
-    $sql = "SELECT * FROM tbl_documentrequest WHERE 1=1";
-    
-    // Add conditions based on the provided filters
-    if ($day) {
-        $sql .= " AND DAY(date_request) = ?";
-    }
-    if ($month) {
-        $sql .= " AND MONTH(date_request) = ?";
-    }
-    if ($year) {
-        $sql .= " AND YEAR(date_request) = ?";
-    }
-    
-    // Prepare the statement
-    $stmt = $this->conn->prepare($sql);
+		// Start the base SQL query
+		$sql = "SELECT * FROM tbl_documentrequest WHERE 1=1";
 
-    // Bind parameters dynamically
-    $params = [];
-    if ($day) $params[] = $day;
-    if ($month) $params[] = $month;
-    if ($year) $params[] = $year;
-    
-    // Execute with parameters
-    $stmt->execute($params);
+		// Add conditions based on the provided filters
+		if ($day) {
+			$sql .= " AND DAY(date_request) = ?";
+		}
+		if ($month) {
+			$sql .= " AND MONTH(date_request) = ?";
+		}
+		if ($year) {
+			$sql .= " AND YEAR(date_request) = ?";
+		}
 
-    // Fetch the result set
-    $result = $stmt->get_result();
-    $data = array();
-    while ($row = $result->fetch_assoc()) {
-        $data[] = $row;
-    }
-    return $data;
+		// Prepare the statement
+		$stmt = $this->conn->prepare($sql);
+
+		// Bind parameters dynamically
+		$params = [];
+		if ($day) $params[] = $day;
+		if ($month) $params[] = $month;
+		if ($year) $params[] = $year;
+
+		// Execute with parameters
+		$stmt->execute($params);
+
+		// Fetch the result set
+		$result = $stmt->get_result();
+		$data = array();
+		while ($row = $result->fetch_assoc()) {
+			$data[] = $row;
+		}
+		return $data;
 	}
 
 	public function print_documentrequest()
@@ -614,10 +614,10 @@ class class_model
 		}
 	}
 
-	public function add_user($complete_name, $desgination, $email_address, $phone_number, $username, $password, $status)
+	public function add_user($complete_name, $desgination, $email_address, $phone_number, $username, $password, $status, $role)
 	{
-		$stmt = $this->conn->prepare("INSERT INTO `tbl_usermanagement` (`complete_name`, `desgination`, `email_address`, `phone_number`, `username`, `password`, `status`) VALUES(?, ?, ?, ?, ?, ?, ?)") or die($this->conn->error);
-		$stmt->bind_param("sssssss", $complete_name, $desgination, $email_address, $phone_number, $username, $password, $status);
+		$stmt = $this->conn->prepare("INSERT INTO `tbl_usermanagement` (`complete_name`, `desgination`, `email_address`, `phone_number`, `username`, `password`, `status`, `role`) VALUES(?, ?, ?, ?, ?, ?, ?, ?)") or die($this->conn->error);
+		$stmt->bind_param("ssssssss", $complete_name, $desgination, $email_address, $phone_number, $username, $password, $status, $role);
 		if ($stmt->execute()) {
 			$stmt->close();
 			$this->conn->close();
@@ -625,10 +625,10 @@ class class_model
 		}
 	}
 
-	public function add_admin_treasurer($complete_name, $desgination, $email_address, $phone_number, $username, $password, $status)
+	public function add_admin_treasurer($complete_name, $desgination, $email_address, $phone_number, $username, $password, $status, $role)
 	{
-		$stmt = $this->conn->prepare("INSERT INTO `tbl_usermanagement` (`complete_name`, `desgination`, `email_address`, `phone_number`, `username`, `password`, `status`) VALUES(?, ?, ?, ?, ?, ?, ?)") or die($this->conn->error);
-		$stmt->bind_param("sssssss", $complete_name, $desgination, $email_address, $phone_number, $username, $password, $status);
+		$stmt = $this->conn->prepare("INSERT INTO `tbl_usermanagement` (`complete_name`, `desgination`, `email_address`, `phone_number`, `username`, `password`, `status` , `role`) VALUES(?, ?, ?, ?, ?, ?, ?, ?)") or die($this->conn->error);
+		$stmt->bind_param("ssssssss", $complete_name, $desgination, $email_address, $phone_number, $username, $password, $status, $role);
 		if ($stmt->execute()) {
 			$stmt->close();
 			$this->conn->close();
@@ -636,10 +636,10 @@ class class_model
 		}
 	}
 
-	public function add_admin_library($complete_name, $desgination, $email_address, $phone_number, $username, $password, $status)
+	public function add_admin_library($complete_name, $desgination, $email_address, $phone_number, $username, $password, $status, $role)
 	{
-		$stmt = $this->conn->prepare("INSERT INTO `tbl_usermanagement` (`complete_name`, `desgination`, `email_address`, `phone_number`, `username`, `password`, `status`) VALUES(?, ?, ?, ?, ?, ?, ?)") or die($this->conn->error);
-		$stmt->bind_param("sssssss", $complete_name, $desgination, $email_address, $phone_number, $username, $password, $status);
+		$stmt = $this->conn->prepare("INSERT INTO `tbl_usermanagement` (`complete_name`, `desgination`, `email_address`, `phone_number`, `username`, `password`, `status` , `role`) VALUES(?, ?, ?, ?, ?, ?, ?, ?)") or die($this->conn->error);
+		$stmt->bind_param("ssssssss", $complete_name, $desgination, $email_address, $phone_number, $username, $password, $status, $role);
 		if ($stmt->execute()) {
 			$stmt->close();
 			$this->conn->close();
@@ -647,10 +647,10 @@ class class_model
 		}
 	}
 
-	public function add_admin_custudian($complete_name, $desgination, $email_address, $phone_number, $username, $password, $status)
+	public function add_admin_custodian($complete_name, $desgination, $email_address, $phone_number, $username, $password, $status, $role)
 	{
-		$stmt = $this->conn->prepare("INSERT INTO `tbl_usermanagement` (`complete_name`, `desgination`, `email_address`, `phone_number`, `username`, `password`, `status`) VALUES(?, ?, ?, ?, ?, ?, ?)") or die($this->conn->error);
-		$stmt->bind_param("sssssss", $complete_name, $desgination, $email_address, $phone_number, $username, $password, $status);
+		$stmt = $this->conn->prepare("INSERT INTO `tbl_usermanagement` (`complete_name`, `desgination`, `email_address`, `phone_number`, `username`, `password`, `status`, `role`) VALUES(?, ?, ?, ?, ?, ?, ?, ?)") or die($this->conn->error);
+		$stmt->bind_param("ssssssss", $complete_name, $desgination, $email_address, $phone_number, $username, $password, $status, $role);
 		if ($stmt->execute()) {
 			$stmt->close();
 			$this->conn->close();
@@ -658,10 +658,10 @@ class class_model
 		}
 	}
 
-	public function add_admin_dean($complete_name, $desgination, $email_address, $phone_number, $username, $password, $status)
+	public function add_admin_dean($complete_name, $desgination, $email_address, $phone_number, $username, $password, $status, $role)
 	{
-		$stmt = $this->conn->prepare("INSERT INTO `tbl_usermanagement` (`complete_name`, `desgination`, `email_address`, `phone_number`, `username`, `password`, `status`) VALUES(?, ?, ?, ?, ?, ?, ?)") or die($this->conn->error);
-		$stmt->bind_param("sssssss", $complete_name, $desgination, $email_address, $phone_number, $username, $password, $status);
+		$stmt = $this->conn->prepare("INSERT INTO `tbl_usermanagement` (`complete_name`, `desgination`, `email_address`, `phone_number`, `username`, `password`, `status`, `role`) VALUES(?, ?, ?, ?, ?, ?, ?, ?)") or die($this->conn->error);
+		$stmt->bind_param("ssssssss", $complete_name, $desgination, $email_address, $phone_number, $username, $password, $status, $role);
 		if ($stmt->execute()) {
 			$stmt->close();
 			$this->conn->close();
@@ -861,28 +861,25 @@ class class_model
 	}
 
 	public function fetchMonthly_documentrequest($year = null)
-{
-    $sql = "SELECT MONTH(date_request) as month, COUNT(*) as count FROM tbl_documentrequest WHERE 1=1";
-    if ($year) {
-        $sql .= " AND YEAR(date_request) = ?";
-    }
-    $sql .= " GROUP BY MONTH(date_request)";
-    
-    $stmt = $this->conn->prepare($sql);
-    $params = [];
-    if ($year) $params[] = $year;
-    
-    $stmt->execute($params);
-    $result = $stmt->get_result();
-    $data = array();
-    while ($row = $result->fetch_assoc()) {
-        $data[] = $row;
-    }
-    return $data;
-}
-	
+	{
+		$sql = "SELECT MONTH(date_request) as month, COUNT(*) as count FROM tbl_documentrequest WHERE 1=1";
+		if ($year) {
+			$sql .= " AND YEAR(date_request) = ?";
+		}
+		$sql .= " GROUP BY MONTH(date_request)";
 
-	
+		$stmt = $this->conn->prepare($sql);
+		$params = [];
+		if ($year) $params[] = $year;
+
+		$stmt->execute($params);
+		$result = $stmt->get_result();
+		$data = array();
+		while ($row = $result->fetch_assoc()) {
+			$data[] = $row;
+		}
+		return $data;
+	}
 }
 
 ?>
